Add files via upload
Reflection

My reflection on industrial visit and talk at Petronas and Huawei company is that In a world that is changing quickly, I believe it is critical for me to be proactive and forward thinking. It is important in staying one step ahead of other people in this highly competitive setting. The rapid development of technology serves a motivation to always be up
to date in an ever-changing environment. It brings impact by being more adaptable to the fast-evolving technology.