Assignment 3 Video (UTM Digital)
Reflection

These whole activities really benefit us all as students. For example, we have been exposed to the place where the UTM Digital staff work, which is called Open Office. At the Open Office, we learn what work the staff need to do and how they relieve their tension in leisure time and more. These activities make us realize that the university is not just a place for us to learn and then take the exam only, but we also need to gain more experience as much as possible, as much experience of how the industry actually works. For example, we need to improve on how to interact with the client or customer and how to make a professional presentation and much more. All UTM Digital staff are very helpful and also provide very great details on when or about what UTM Digital is all about. Some of them are very generous to show us their smart meeting room. They have a smart TV to make it easier to show their slides. From our experience, we were able to try that smart TV and, for your information, this TV is able to take a photo of us. That is all from us. Hopefully we can contribute more to the industry. There is more advanced technology in the future that is waiting to be created.